refactor(aggregated-assets): tighten the package surface (LIVE-35226)
Review feedback from @ysitbon on #20345.

- market.ts moves to internals/ and leaves the barrel. Nothing outside the
  package imported it: every external consumer of dadaIdToMarketId and
  PartialMarketItemResponse resolves the legacy live-common declaration, not
  this one. The two copies still need de-duplicating in the retarget tasks.
- the five single-function pure helpers become internals/utils.ts, with their
  tests merged. collectAllByCategory stays its own file: it does network I/O
  and paginates, which is not a util.
- ONE_DAY_IN_SECONDS leaves types.ts for constants.ts, written as 24 * 60 * 60.
  It stays exported because the live-common shim re-exports it.
- the mocks move under src/fixtures/. The .mock.ts suffix is kept so knip's
  '**/*.mock.*' ignore still applies, and the ./mock* export names are
  unchanged, so no consumer moves.
