+++ domain/api/aggregated-assets/src/constants.ts
@@ -1,0 +1,2 @@
+/** Cache lifetime for the collections DADA only changes daily (`keepUnusedDataFor`, in seconds). */
+export const ONE_DAY_IN_SECONDS = 24 * 60 * 60;
