+++ domain/api/aggregated-assets/src/fixtures/assetsData.mock.ts
