+++ domain/api/aggregated-assets/src/fixtures/stablecoins.mock.ts
