+++ domain/api/aggregated-assets/src/fixtures/stocks.mock.ts
