+++ domain/api/aggregated-assets/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./constants";
 export * from "./schema";
 export * from "./types";
 export * from "./transforms";
@@ -5,5 +6,4 @@ export * from "./requests";
 export * from "./accessors";
 export * from "./api";
 export * from "./errors";
-export * from "./market";
 export * from "./pagination";
